"Update ImageUpload page and add logout functionality to user navbar"
This commit updates the ImageUpload page to include a form for submitting an image, as well as a preview of the selected image. It also adds a logout function to the user navbar, which sends a GET request to the server to log out the user.

Additionally, the commit updates the admin navbar to include a link to the Image page, and updates the user navbar to include a link to the login page when the user is logged out. The commit also includes some minor changes to the styling of the navbar components.