"Added conditional rendering for message when images array is empty in Gallery page"
This commit adds a new div element that displays the text "Please check after a few days..." if the images array is empty. This is achieved by adding a conditional (ternary) operator in the return statement of the Gallery component. The existing div with the grid layout is wrapped in a new div with a closing tag > added to its end.